feat: [BBND-1498] move dividend tests from Equity to Dividend
